backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
asyncpg
psycopg[binary]
google-generativeai